add ring buffer module
